# Conflicts:
#	internal/server/handlers/updateMetrics.go

+++ internal/server/storage/dbStorage.go
@@ -9,7 +9,7 @@ func createTables(db *sql.DB) error {
 	sqlStatement := `
 CREATE TABLE IF NOT EXISTS counter (
     name VARCHAR(64),
-    value NUMERIC,
+    value integer,
     PRIMARY KEY (name)
 );
 CREATE TABLE IF NOT EXISTS gauge (
